+++ packages/renderer/tailwind.config.cjs
@@ -1,1 +1,1 @@
-../../tailwind.config.cjs
+module.exports = require('../../tailwind.config.cjs');