+++ packages/adapter-evm/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @openzeppelin/transaction-form-adapter-evm
 
+## 0.5.0
+
+### Minor Changes
+
+- [#106](https://github.com/OpenZeppelin/contracts-ui-builder/pull/106) [`ce96c10`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/ce96c104e9e5df22ba335a8746cda740a70dbd0b) Thanks [@pasevin](https://github.com/pasevin)! - Resolve legacy OpenZeppelin proxy implementation/admin via storage slots; expose adminAddress in proxy info; delegate auto-query filtering to adapter to avoid admin-only getters; add storage-slot debug logs.
+
+### Patch Changes
+
+- Updated dependencies [[`ce96c10`](https://github.com/OpenZeppelin/contracts-ui-builder/commit/ce96c104e9e5df22ba335a8746cda740a70dbd0b)]:
+  - @openzeppelin/contracts-ui-builder-types@0.4.0
+  - @openzeppelin/contracts-ui-builder-react-core@0.2.5
+  - @openzeppelin/contracts-ui-builder-ui@0.5.1
+  - @openzeppelin/contracts-ui-builder-utils@0.4.1
+
 ## 0.4.0
 
 ### Minor Changes